CS fixes
