+++ test/ProgressBarTest.php
@@ -9,6 +9,9 @@
 
 namespace ZendTest\ProgressBar;
 
+use ZendTest\ProgressBar\TestAsset\MockUp;
+use ZendTest\ProgressBar\TestAsset\Stub;
+
 /**
  * @group      Zend_ProgressBar
  */
@@ -84,100 +87,10 @@ class ProgressBarTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(null, $progressBar->getTimeRemaining());
     }
 
+    // @codingStandardsIgnoreStart
     protected function _getProgressBar($min, $max, $persistenceNamespace = null)
     {
+        // @codingStandardsIgnoreEnd
         return new Stub(new MockUp(), $min, $max, $persistenceNamespace);
-    }
-}
-
-class Stub extends \Zend\ProgressBar\ProgressBar
-{
-    public function sleep($seconds)
-    {
-        $this->startTime -= $seconds;
-    }
-
-    public function getCurrent()
-    {
-        return $this->adapter->getCurrent();
-    }
-
-    public function getMax()
-    {
-        return $this->adapter->getMax();
-    }
-
-    public function getPercent()
-    {
-        return $this->adapter->getPercent();
-    }
-
-    public function getTimeTaken()
-    {
-        return $this->adapter->getTimeTaken();
-    }
-
-    public function getTimeRemaining()
-    {
-        return $this->adapter->getTimeRemaining();
-    }
-
-    public function getText()
-    {
-        return $this->adapter->getText();
-    }
-}
-
-class MockUp extends \Zend\ProgressBar\Adapter\AbstractAdapter
-{
-    protected $_current;
-    protected $_max;
-    protected $_percent;
-    protected $_timeTaken;
-    protected $_timeRemaining;
-    protected $_text;
-
-    public function notify($current, $max, $percent, $timeTaken, $timeRemaining, $text)
-    {
-        $this->_current       = $current;
-        $this->_max           = $max;
-        $this->_percent       = $percent;
-        $this->_timeTaken     = $timeTaken;
-        $this->_timeRemaining = $timeRemaining;
-        $this->_text          = $text;
-    }
-
-    public function finish()
-    {
-    }
-
-    public function getCurrent()
-    {
-        return $this->_current;
-    }
-
-    public function getMax()
-    {
-        return $this->_max;
-    }
-
-    public function getPercent()
-    {
-        return $this->_percent;
-    }
-
-    public function getTimeTaken()
-    {
-        return $this->_timeTaken;
-    }
-
-    public function getTimeRemaining()
-    {
-        return $this->_timeRemaining;
-    }
-
-    public function getText()
-    {
-        return $this->_text;
     }
 }
